requirements.txt
streamlit>=1.32.0
google-genai>=1.0.0
opik>=0.1.0
python-dotenv>=1.0.0
yfinance>=0.2.0
requests>=2.31.0
pandas>=2.0.0
plotly>=5.18.0